requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
numpy==1.16.3
pandas==0.24.2
pypinyin==0.35.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.1
